Move beerjson types to the types folder
